+++ src/internal/observable/zip.ts
@@ -3,7 +3,7 @@ import { ObservableInputTuple } from '../types';
 import { innerFrom } from './innerFrom';
 import { argsOrArgArray } from '../util/argsOrArgArray';
 import { EMPTY } from './empty';
-import { OperatorSubscriber } from '../operators/OperatorSubscriber';
+import { createOperatorSubscriber } from '../operators/OperatorSubscriber';
 import { popResultSelector } from '../util/args';
 
 export function zip<A extends readonly unknown[]>(sources: [...ObservableInputTuple<A>]): Observable<A>;
@@ -74,7 +74,7 @@ export function zip(...args: unknown[]): Observable<unknown> {
         // access the related buffers and completion properties
         for (let sourceIndex = 0; !subscriber.closed && sourceIndex < sources.length; sourceIndex++) {
           innerFrom(sources[sourceIndex]).subscribe(
-            new OperatorSubscriber(
+            createOperatorSubscriber(
               subscriber,
               (value) => {
                 buffers[sourceIndex].push(value);