+++ src/internal/operators/distinct.ts
@@ -1,7 +1,7 @@
 import { Observable } from '../Observable';
 import { MonoTypeOperatorFunction } from '../types';
 import { operate } from '../util/lift';
-import { OperatorSubscriber } from './OperatorSubscriber';
+import { createOperatorSubscriber } from './OperatorSubscriber';
 import { noop } from '../util/noop';
 
 /**
@@ -65,7 +65,7 @@ export function distinct<T, K>(keySelector?: (value: T) => K, flushes?: Observab
   return operate((source, subscriber) => {
     const distinctKeys = new Set();
     source.subscribe(
-      new OperatorSubscriber(subscriber, (value) => {
+      createOperatorSubscriber(subscriber, (value) => {
         const key = keySelector ? keySelector(value) : value;
         if (!distinctKeys.has(key)) {
           distinctKeys.add(key);
@@ -74,6 +74,6 @@ export function distinct<T, K>(keySelector?: (value: T) => K, flushes?: Observab
       })
     );
 
-    flushes?.subscribe(new OperatorSubscriber(subscriber, () => distinctKeys.clear(), noop));
+    flushes?.subscribe(createOperatorSubscriber(subscriber, () => distinctKeys.clear(), noop));
   });
 }