+++ src/internal/operators/every.ts
@@ -1,7 +1,7 @@
 import { Observable } from '../Observable';
 import { Falsy, OperatorFunction } from '../types';
 import { operate } from '../util/lift';
-import { OperatorSubscriber } from './OperatorSubscriber';
+import { createOperatorSubscriber } from './OperatorSubscriber';
 
 export function every<T>(predicate: BooleanConstructor): OperatorFunction<T, Exclude<T, Falsy> extends never ? false : boolean>;
 /** @deprecated Use a closure instead of a `thisArg`. Signatures accepting a `thisArg` will be removed in v8. */
@@ -48,7 +48,7 @@ export function every<T>(
   return operate((source, subscriber) => {
     let index = 0;
     source.subscribe(
-      new OperatorSubscriber(
+      createOperatorSubscriber(
         subscriber,
         (value) => {
           if (!predicate.call(thisArg, value, index++, source)) {