+++ src/internal/operators/skipUntil.ts
@@ -1,7 +1,7 @@
 import { Observable } from '../Observable';
 import { MonoTypeOperatorFunction } from '../types';
 import { operate } from '../util/lift';
-import { OperatorSubscriber } from './OperatorSubscriber';
+import { createOperatorSubscriber } from './OperatorSubscriber';
 import { innerFrom } from '../observable/innerFrom';
 import { noop } from '../util/noop';
 
@@ -51,7 +51,7 @@ export function skipUntil<T>(notifier: Observable<any>): MonoTypeOperatorFunctio
   return operate((source, subscriber) => {
     let taking = false;
 
-    const skipSubscriber = new OperatorSubscriber(
+    const skipSubscriber = createOperatorSubscriber(
       subscriber,
       () => {
         skipSubscriber?.unsubscribe();
@@ -62,6 +62,6 @@ export function skipUntil<T>(notifier: Observable<any>): MonoTypeOperatorFunctio
 
     innerFrom(notifier).subscribe(skipSubscriber);
 
-    source.subscribe(new OperatorSubscriber(subscriber, (value) => taking && subscriber.next(value)));
+    source.subscribe(createOperatorSubscriber(subscriber, (value) => taking && subscriber.next(value)));
   });
 }